+++ tests/TestCase/Model/Entity/PhotoTest.php
@@ -60,6 +60,16 @@ class PhotoTest extends EntityTestCase
     }
 
     /**
+     * Test for `_getDescription()` method
+     * @test
+     */
+    public function testDescriptionAccessors()
+    {
+        $this->assertNotNull($this->Entity->get('description'));
+        $this->assertSame('', $this->Entity->set('description', null)->get('description'));
+    }
+
+    /**
      * Test for `_getPlainDescription()` method
      * @test
      */