+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 * fixed `TestSuite` classes, improved to use with other plugins;
 * fixed old bug on loading theme plugins;
 * fixed some template files, improved to use with other plugins;
+* fixed little bug for the `description` field for `Banner` and `Photo` entities;
 * updated for `php-tools` 1.4.5;
 * removed redundant code in routes;
 * added `phpstan`, so fixed some code and descriptions.